Source/Chatbot/PhoBERT/requirements.txt
flask
torch
transformers
h5py
numpy
scikit-learn
aiohttp
requests
python-dotenv
